requirements.txt
openai
python-dotenv
streamlit
flake8
bandit
gitpython
pytest
